+++ src/main/java/com/evanbuss/webscraper/ui/controllers/HTMLTabController.java
@@ -1,0 +1,14 @@
+package com.evanbuss.webscraper.ui.controllers;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.TextArea;
+
+public class HTMLTabController {
+  @FXML
+  TextArea siteDataTextArea;
+  private MainController mainController;
+
+  void inject(MainController mainController) {
+    this.mainController = mainController;
+  }
+}
