--- src/main/java/com/evanbuss/webscraper/ui/controllers/SelectorController.java
@@ -1,4 +1,0 @@
-package com.evanbuss.webscraper.ui.controllers;
-
-public class SelectorController {
-}
