+++ src/main/java/com/evanbuss/webscraper/ui/controllers/SelectorsTabController.java
@@ -1,0 +1,10 @@
+package com.evanbuss.webscraper.ui.controllers;
+
+public class SelectorsTabController {
+
+  private MainController mainController;
+
+  void inject(MainController mainController) {
+    this.mainController = mainController;
+  }
+}
